+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # 1.0.0 (1 January, 2020)
 
 - Initial release
+
+# 1.1.4 (19 May, 2020)
+
+- Bug fix 
+    - after auth will always error
+- Add a bit more Error Handling
+- new config `forceAuth`
+    - not throw error if auth failed
+- More comments